距離行列を整数に修正

+++ VRP/cpp/src/construct.cpp
@@ -9,7 +9,7 @@
 
 using namespace std;
 
-vector<vector<int>> InsertConstruct(const vector<vector<float>>& dis_mat,const vector<int>& weights,
+vector<vector<int>> InsertConstruct(const vector<vector<int>>& dis_mat,const vector<int>& weights,
                                     int truck_capacity,int truck_size,vector<pair<int,int>>& truck_ids,
                                     bool construct_randomly){
     vector<vector<int>> orders(truck_size,vector<int>(2,0));
