+++ VRP/cpp/src/construct.hpp
@@ -5,6 +5,6 @@
 using std::vector;
 using std::pair;
 
-vector<vector<int>> InsertConstruct(const vector<vector<float>>& dis_mat,const vector<int>& weights,
+vector<vector<int>> InsertConstruct(const vector<vector<int>>& dis_mat,const vector<int>& weights,
                                     int truck_capacity,int truck_size,vector<pair<int,int>>& truck_ids,
                                     bool construct_randomly=false);
