+++ VRP/cpp/src/util.hpp
@@ -8,11 +8,11 @@ using std::vector;
 using std::pair;
 using std::set;
 
-double TotalDistance(const vector<vector<int>>& orders,
-                        const vector<vector<float>>& dis_mat);
+int TotalDistance(const vector<vector<int>>& orders,
+                        const vector<vector<int>>& dis_mat);
 
-double TotalDistance(const vector<int>& order,
-                    const vector<vector<float>>& dis_mat);
+int TotalDistance(const vector<int>& order,
+                    const vector<vector<int>>& dis_mat);
 
 int TotalWeight(const vector<int>& order,const vector<int>& weights);
 
@@ -27,9 +27,9 @@ void ShowOrdersInfo(const vector<vector<int>>& orders,const vector<int>& weights
 
 vector<vector<int>> comb(int n, int r);
 
-vector<vector<int>> ConstructNeighborList(const int n,const vector<vector<float>>& dis_mat);
+vector<vector<int>> ConstructNeighborList(const int n,const vector<vector<int>>& dis_mat);
 
-vector<set<int>> ConstructNNList(const vector<vector<float>>& dis_mat,
+vector<set<int>> ConstructNNList(const vector<vector<int>>& dis_mat,
                                 const vector<vector<int>>& neighbor_list);
 
 void UpdateTruckIds(const vector<vector<int>>& orders,vector<pair<int,int>>& truck_ids);
@@ -40,11 +40,11 @@ void ShowTruckIds(const vector<pair<int,int>>& truck_ids);
 
 namespace util{
     void ShowTwoOptStarDiffs(const vector<int>& weights,vector<vector<int>>& orders,
-                const vector<vector<float>>& dis_mat,const int truck_capacity,
+                const vector<vector<int>>& dis_mat,const int truck_capacity,
                 vector<pair<int,int>>& truck_ids);
     
     void SubTwoOptStar(const vector<int>& weights,vector<vector<int>>& orders,
-                const vector<vector<float>>& dis_mat,const int truck_capacity,
+                const vector<vector<int>>& dis_mat,const int truck_capacity,
                 const int i,const int j,vector<pair<int,int>>& truck_ids);
     
     void ShowDifInfos(int i,int j,double change_cost_dif,
@@ -53,11 +53,11 @@ namespace util{
                         bool isSimpled=false);
 
     void ShowFastTwoOptStarDiffs(const vector<int>& weights,vector<vector<int>>& orders,
-                const vector<vector<float>>& dis_mat,const int truck_capacity,
+                const vector<vector<int>>& dis_mat,const int truck_capacity,
                 vector<pair<int,int>>& truck_ids,const vector<set<int>>& nn_list);
     
     void SubFastTwoOptStar(const vector<int>& weights,vector<vector<int>>& orders,
-                    const vector<vector<float>>& dis_mat,const int truck_capacity,
+                    const vector<vector<int>>& dis_mat,const int truck_capacity,
                     vector<pair<int,int>>& truck_ids,const vector<set<int>>& nn_list);
 }
